Add delete employee confirmation dialog

+++ src/app/components/employee-details/employee-details.component.ts
@@ -73,8 +73,14 @@ export class EmployeeDetailsComponent implements OnInit {
   }
 
   onDelete() {
-    this.employeeService.deleteEmployee(this.employeeId());
-    this.router.navigate(['/'], { replaceUrl: true });
+    const confirmation = window.confirm(
+      'Are you sure that you want to delete ' + this.employee?.name + '?'
+    );
+
+    if (confirmation) {
+      this.employeeService.deleteEmployee(this.employeeId());
+      this.router.navigate(['/'], { replaceUrl: true });
+    }
   }
 
   onToggleModal() {
